Split up file into a file per class
