+++ addons/geoscript/geoscript.py
@@ -1,284 +1,7 @@
 #!/usr/bin/python3
 import bpy
 
-class AbstractSocket():
-    """Any type of output inside a node tree. This class handles the positioning and routing
-    of new nodes in a node tree."""
-    
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
-        if not node_tree:
-            node_tree = bpy.types.GeometryNodeTree()
-            
-        self.node_tree = node_tree
-        self.socket_reference = socket_reference
-        self.layer = layer
-    
-    @staticmethod
-    def new_node(input_list, node_type: str = ''):
-        """Add a new node to the right of all input sockets."""
-        
-        # First calculate which is the rightmost layer of the input sockets:
-        max_layer = 0
-        for socket in input_list:
-            if not isinstance(socket, AbstractSocket):
-                raise TypeError("Socket inside socket list is not an instance of AbstractSocket, but of {}. This is a bug, please report to the developers.".format(socket.__class__))
-            
-            max_layer = max(max_layer, socket.layer)
-        
-        new_layer = max_layer + 1
-        
-        # Then create a new node to the right of that rightmost layer:
-        new_node = input_list[0].node_tree.nodes.new(node_type)
-        new_node.location = (200.0 * new_layer, 0.0)
-        
-        return (new_node, new_layer)
-        
-
-class AbstractTensor(AbstractSocket):
-    """A mathematical object on which various operations can be performed, such as a Scalar or a Vector."""
-    
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
-        super().__init__(node_tree, socket_reference, layer)
-    
-    @staticmethod
-    def math_operation_unary(input, operation: str = 'ADD', use_clamp: bool = False):
-        return NotImplemented
-    
-    @staticmethod
-    def math_operation_binary(left, right, operation: str = 'ADD', use_clamp: bool = False):
-        return NotImplemented
-    
-    # Add:
-    def __add__(self, other):
-        return self.math_operation_binary(self, other, operation = 'ADD', use_clamp = False);
-    
-    def __radd__(self, other):
-        return self.math_operation_binary(other, self, operation = 'ADD', use_clamp = False);
-    
-    # Subtract:
-    def __sub__(self, other):
-        return self.math_operation_binary(self, other, operation = 'SUBTRACT', use_clamp = False);
-    
-    def __rsub__(self, other):
-        return self.math_operation_binary(other, self, operation = 'SUBTRACT', use_clamp = False);
-    
-    # Multiply:
-    def __mul__(self, other):
-        return self.math_operation_binary(self, other, operation = 'MULTIPLY', use_clamp = False);
-    
-    def __rmul__(self, other):
-        return self.math_operation_binary(other, self, operation = 'MULTIPLY', use_clamp = False);
-    
-    # Divide:
-    def __div__(self, other):
-        return self.math_operation_binary(self, other, operation = 'DIVIDE', use_clamp = False);
-    
-    def __rdiv__(self, other):
-        return self.math_operation_binary(other, self, operation = 'DIVIDE', use_clamp = False);
-    
-    def __truediv__(self, other):
-        return self.math_operation_binary(self, other, operation = 'DIVIDE', use_clamp = False);
-    
-    def __rtruediv__(self, other):
-        return self.math_operation_binary(other, self, operation = 'DIVIDE', use_clamp = False);
-    
-    # Modulo:
-    def __mod__(self, other):
-        return self.math_operation_binary(self, other, operation = 'MODULO', use_clamp = False);
-    
-    def __rmod__(self, other):
-        return self.math_operation_binary(other, self, operation = 'MODULO', use_clamp = False);
-    
-    # Power:
-    def __pow__(self, other):
-        return self.math_operation_binary(self, other, operation = 'POWER', use_clamp = False);
-    
-    def __rpow__(self, other):
-        return self.math_operation_binary(other, self, operation = 'POWER', use_clamp = False);
-    
-    # Unary operations:
-    def __abs__(self):
-        return self.math_operation_unary(self, other, operation = 'ABSOLUTE', use_clamp = False);
-    
-    def __neg__(self):
-        return self.math_operation_binary(-1.0, self, operation = 'MULTIPLY', use_clamp = False);
-    
-    def __round__(self):
-        return self.math_operation_unary(self, operation = 'ROUND', use_clamp = False);
-    
-    def __trunc__(self):
-        return self.math_operation_unary(self, operation = 'TRUNC', use_clamp = False);
-    
-    def __floor__(self):
-        return self.math_operation_unary(self, operation = 'FLOOR', use_clamp = False);
-    
-    def __ceil__(self):
-        return self.math_operation_unary(self, operation = 'CEIL', use_clamp = False);
-
-
-class Vector(AbstractTensor):
-    """A mathematics operation in a Geometry Node tree. Maps to a "Math" node."""
-    
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
-        super().__init__(node_tree, socket_reference, layer)
-    
-    @staticmethod
-    def math_operation_unary(input, operation: str = 'ADD', use_clamp: bool = False):
-        #math_node = input.node_tree.nodes.new('ShaderNodeVectorMath')
-        math_node, layer = input.new_node([input], 'ShaderNodeVectorMath')
-        math_node.operation = operation
-        
-        input.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-        
-        return Vector(input.node_tree, math_node.outputs[0], layer)
-    
-    @staticmethod
-    def math_operation_binary(left, right, operation: str = 'ADD', use_clamp: bool = False):
-        if isinstance(right, left.__class__):
-            #math_node = left.node_tree.nodes.new('ShaderNodeVectorMath')
-            math_node, layer = left.new_node([left, right], 'ShaderNodeVectorMath')
-            math_node.operation = operation
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            left.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Vector(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(right, float):
-            #math_node = left.node_tree.nodes.new('ShaderNodeVectorMath')
-            math_node, layer = left.new_node([left], 'ShaderNodeVectorMath')
-            math_node.operation = operation
-            math_node.inputs[1].default_value = right
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            
-            return Vector(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(left, float):
-            #math_node = right.node_tree.nodes.new('ShaderNodeVectorMath')
-            math_node, layer = right.new_node([right], 'ShaderNodeVectorMath')
-            math_node.operation = operation
-            math_node.inputs[0].default_value = left
-            
-            right.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Vector(right.node_tree, math_node.outputs[0], layer)
-        
-        else:
-            return NotImplemented
-
-
-class Scalar(AbstractTensor):
-    """A mathematics operation in a Geometry Node tree. Maps to a "Math" node."""
-    
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
-        super().__init__(node_tree, socket_reference, layer)
-    
-    @staticmethod
-    def math_operation_unary(input, operation: str = 'ADD', use_clamp: bool = False):
-        #math_node = input.node_tree.nodes.new('ShaderNodeMath')
-        math_node, layer = input.new_node([input], 'ShaderNodeMath')
-        math_node.operation = operation
-        math_node.use_clamp = use_clamp
-        
-        input.node_tree.links.new(left.socket_reference, math_node.inputs[0], layer)
-        
-        return Scalar(input.node_tree, math_node.outputs[0], layer)
-    
-    @staticmethod
-    def math_operation_binary(left, right, operation: str = 'ADD', use_clamp: bool = False):
-        if isinstance(right, left.__class__):
-            #math_node = left.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = left.new_node([left, right], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            left.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Scalar(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(right, float):
-            #math_node = left.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = left.new_node([left], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            math_node.inputs[1].default_value = right
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            
-            return Scalar(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(left, float):
-            #math_node = right.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = right.new_node([right], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            math_node.inputs[0].default_value = left
-            
-            right.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Scalar(right.node_tree, math_node.outputs[0], layer)
-        
-        else:
-            return NotImplemented
-
-
-class Scalar(AbstractTensor):
-    """A mathematics operation in a Geometry Node tree. Maps to a "Math" node."""
-    
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
-        super().__init__(node_tree, socket_reference, layer)
-    
-    @staticmethod
-    def math_operation_unary(input, operation: str = 'ADD', use_clamp: bool = False):
-        #math_node = input.node_tree.nodes.new('ShaderNodeMath')
-        math_node, layer = input.new_node([input], 'ShaderNodeMath')
-        math_node.operation = operation
-        math_node.use_clamp = use_clamp
-        
-        input.node_tree.links.new(left.socket_reference, math_node.inputs[0], layer)
-        
-        return Scalar(input.node_tree, math_node.outputs[0], layer)
-    
-    @staticmethod
-    def math_operation_binary(left, right, operation: str = 'ADD', use_clamp: bool = False):
-        if isinstance(right, left.__class__):
-            #math_node = left.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = left.new_node([left, right], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            left.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Scalar(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(right, float):
-            #math_node = left.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = left.new_node([left], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            math_node.inputs[1].default_value = right
-            
-            left.node_tree.links.new(left.socket_reference, math_node.inputs[0])
-            
-            return Scalar(left.node_tree, math_node.outputs[0], layer)
-        
-        elif isinstance(left, float):
-            #math_node = right.node_tree.nodes.new('ShaderNodeMath')
-            math_node, layer = right.new_node([right], 'ShaderNodeMath')
-            math_node.operation = operation
-            math_node.use_clamp = use_clamp
-            math_node.inputs[0].default_value = left
-            
-            right.node_tree.links.new(right.socket_reference, math_node.inputs[1])
-            
-            return Scalar(right.node_tree, math_node.outputs[0], layer)
-        
-        else:
-            return NotImplemented
-
+from .types import *
 
 class GeometryFunction():
     """Generate geometry node trees. Corresponds to a Geometry Nodes tree."""
