+++ addons/geoscript/types/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/python3
+
+from .abstract_socket import AbstractSocket
+from .abstract_tensor import AbstractTensor
+from .scalar import Scalar
+from .vector3 import Vector
+
+__all__ = ["AbstractSocket", "AbstractTensor", "Scalar", "Vector"]
+
